trim the fat
