+++ .github/workflows/release-recurring.yml
@@ -43,11 +43,11 @@ jobs:
 
       - name: Bump version
         shell: bash
-        run: BUILD_REF="$(git rev-parse --short HEAD)${{ github.event_name == 'pull_request' && '.pr-$PR_NUMBER' || '' }}" npm run app-bump-version
+        run: BUILD_REF="$(git rev-parse --short HEAD)${{ github.event_name == 'pull_request' && '.pr-$PR_NUMBER' || '' }}" npm run bump-version --prefix packages/insomnia
 
       - name: Package
         shell: bash
-        run: BUILD_TARGETS='${{ matrix.build-targets }}' npm run app-package
+        run: BUILD_TARGETS='${{ matrix.build-targets }}' npm run package --prefix packages/insomnia
 
       - name: Upload artifacts
         uses: actions/upload-artifact@v3
