+++ .github/workflows/release-start.yml
@@ -58,7 +58,7 @@ jobs:
         run: npm run version -- --preid "${{ github.event.inputs.channel }}" --yes ${{ github.event.inputs.channel != 'stable' && 'preminor' || 'minor' }}
 
       - name: Config version
-        run: APP_VERSION="${{ env.RELEASE_VERSION }}" npm run app-bump-version
+        run: APP_VERSION="${{ env.RELEASE_VERSION }}" npm run bump-version --prefix packages/insomnia
 
       # https://github.com/lerna/lerna/issues/2832
       - name: Update package-lock.json files
