+++ package.json
@@ -11,32 +11,18 @@
   },
   "homepage": "https://github.com/kong/insomnia#readme",
   "scripts": {
-    "build": "lerna run build --parallel",
+    "dev": "npm start --prefix packages/insomnia-app",
+    "test": "lerna run test --stream --parallel",
+    "type-check": "lerna run type-check",
+    "hard-reset": "npm run clean && npm run bootstrap && npm run dev",
     "lint": "lerna run lint --stream --no-bail",
     "lint:markdown": "markdownlint-cli2 \"**/*.md\" \"#**/node_modules\"",
-    "lint:fix": "lerna run lint:fix --stream --no-bail",
-    "bootstrap": "npm install && lerna bootstrap && lerna run --stream bootstrap",
-    "version": "lerna version",
-    "publish": "lerna publish from-package",
-    "clean": "lerna run clean --parallel --stream && lerna clean --yes && rimraf node_modules",
-    "test": "lerna run --stream --parallel test",
-    "inso-start": "npm start --prefix packages/insomnia-inso",
-    "inso-package": "npm run build:sr --prefix packages/insomnia && npm run package --prefix packages/insomnia-inso",
-    "inso-package:artifacts": "npm run artifacts --prefix packages/insomnia-inso",
-    "watch:app": "npm run build:main.min.js --prefix packages/insomnia && npm run start:dev-server --prefix packages/insomnia",
-    "app-start": "npm start --prefix packages/insomnia",
-    "app-build": "npm run build --prefix packages/insomnia",
-    "app-package": "npm run build --prefix packages/insomnia && npm run package --prefix packages/insomnia",
-    "app-bump-version": "npm run bump-version --prefix packages/insomnia",
     "preinstall": "node ./scripts/check-version.js",
-    "test:smoke:dev": "npm run test:dev --prefix packages/insomnia-smoke-test",
-    "test:smoke:build": "npm run test:build --prefix packages/insomnia-smoke-test",
-    "test:smoke:package": "npm run test:package --prefix packages/insomnia-smoke-test",
-    "test:smoke:cli": "npm run test:cli --prefix packages/insomnia-smoke-test",
-    "hard-reset": "npm run clean && npm run bootstrap && npm run app-start",
-    "type-check": "lerna run type-check",
-    "changelog-image": "ts-node ./scripts/changelog-image/changelog-image.ts",
-    "dev": "npm start --prefix packages/insomnia"
+    "bootstrap": "npm install && lerna bootstrap && lerna run bootstrap --stream",
+    "bootstrap:electron-builder": "lerna bootstrap --no-ci -- --production --no-optional",
+    "version": "lerna version",
+    "postversion": "lerna clean --yes && lerna bootstrap --no-ci --ignore-scripts -- --package-lock-only --no-audit",
+    "clean": "lerna run clean --parallel --stream && lerna clean --yes && rimraf node_modules"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.5",
