+++ packages/insomnia-common/package.json
@@ -17,12 +17,9 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "bootstrap": "npm run build",
+    "bootstrap": "tsc --build tsconfig.build.json",
     "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "prebuild": "npm run clean",
-    "build": "tsc --build tsconfig.build.json"
+    "clean": "rimraf dist *.tsbuildinfo"
   },
   "devDependencies": {
     "ajv": "^8.6.2",
