+++ packages/insomnia-components/package.json
@@ -23,20 +23,13 @@
     }
   },
   "scripts": {
-    "prebuild": "npm run convert-svg",
+    "test": "jest",
+    "watch": "cross-env ESBUILD_WATCH=true esr esbuild.ts",
     "build:source": "vite build",
     "build:types": "tsc --build tsconfig.build.json",
-    "build": "concurrently --names source,types \"npm run build:source\" \"npm run build:types\"",
-    "clean": "rimraf dist && npm run clean:svg",
-    "clean:svg": "rimraf src/assets/svgr",
-    "clean:svg:index": "rimraf src/assets/svgr/index.tsx",
-    "convert-svg": "npm run clean:svg && svgr --config-file svgr.config.js --out-dir src/assets/svgr src/assets && npm run clean:svg:index",
-    "jest:watch": "jest --watch",
-    "prepare": "npm run convert-svg && npm run build",
-    "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "test": "jest",
-    "watch": "cross-env ESBUILD_WATCH=true esr esbuild.ts"
+    "bootstrap": "svgr --config-file svgr.config.js --out-dir src/assets/svgr src/assets && rimraf src/assets/svgr/index.tsx && npm run build:source && npm run build:types",
+    "clean": "rimraf dist *.tsbuildinfo src/assets/svgr",
+    "lint": "eslint . --ext .js,.ts,.tsx --cache"
   },
   "devDependencies": {
     "@babel/core": "^7.7.7",
@@ -58,7 +51,6 @@
     "@types/testing-library__jest-dom": "^5.9.5",
     "@types/testing-library__react-hooks": "^4.0.0",
     "babel-plugin-inline-react-svg": "^1.1.1",
-    "concurrently": "^7.0.0",
     "cross-env": "^7.0.3",
     "esbuild": "^0.14.27",
     "esbuild-runner": "^2.2.1",
