+++ packages/insomnia-config/package.json
@@ -17,15 +17,10 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "bootstrap": "npm run generate",
+    "bootstrap": "tsc --build tsconfig.build.json && ts-node ./src/generated/generate.ts",
     "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
     "test": "jest",
-    "pregenerate": "npm run build",
-    "generate": "ts-node ./src/generated/generate.ts",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "prebuild": "npm run clean",
-    "build": "tsc --build tsconfig.build.json"
+    "clean": "rimraf dist *.tsbuildinfo"
   },
   "dependencies": {
     "@apideck/better-ajv-errors": "^0.2.6",
