+++ packages/insomnia-inso/package.json
@@ -19,23 +19,14 @@
     "inso": "bin/inso"
   },
   "scripts": {
-    "bootstrap": "npm run build",
-    "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "postclean": "rimraf dist && rimraf binaries && rimraf compressed && rimraf artifacts",
-    "test": "jest --runInBand",
-    "test:watch": "npm run test -- --watch",
-    "test:snapshots": "npm run build && npm run test -- -u",
-    "prebuild": "npm run clean",
-    "build": "webpack --config webpack/webpack.config.development.js",
-    "prebuild:production": "npm run clean",
-    "build:production": "webpack --config webpack/webpack.config.production.js --display errors-only",
     "start": "npm run build -- --watch",
+    "lint": "eslint . --ext .js,.ts,.tsx --cache",
+    "clean": "rimraf dist *.tsbuildinfo binaries compressed artifacts",
+    "test": "jest --runInBand",
+    "build": "webpack --config webpack/webpack.config.development.js",
+    "build:production": "webpack --config webpack/webpack.config.production.js --display errors-only",
     "prepackage": "npm run build:production",
     "package": "ts-node src/scripts/pkg.ts",
-    "pkg": "pkg .",
-    "postpackage": "ts-node src/scripts/verify-pkg.ts",
     "artifacts": "ts-node src/scripts/artifacts.ts"
   },
   "devDependencies": {
