+++ packages/insomnia-smoke-test/package.json
@@ -13,19 +13,14 @@
   },
   "version": "2.13.0-beta.0",
   "scripts": {
-    "bootstrap": "npm run build",
+    "bootstrap": "tsc --build tsconfig.build.json",
     "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "postclean": "rimraf dist",
-    "build": "tsc --build tsconfig.build.json",
+    "clean": "rimraf dist *.tsbuildinfo ",
     "test:dev": "xvfb-maybe cross-env BUNDLE=dev playwright test",
     "test:build": "xvfb-maybe cross-env BUNDLE=build playwright test",
     "test:package": "xvfb-maybe cross-env BUNDLE=package playwright test",
-    "cli": "jest --detectOpenHandles --testPathPattern cli",
     "serve": "ts-node server/index.ts",
-    "with-mock": "concurrently --names server,app --success first --kill-others \"npm run serve\"",
-    "test:cli": "npm run with-mock \"npm run cli\""
+    "test": "concurrently -n server,inso --kill-others \"npm run serve\" \"jest --detectOpenHandles --testPathPattern cli\""
   },
   "devDependencies": {
     "@playwright/test": "^1.20.2",
