+++ packages/insomnia/package.json
@@ -17,26 +17,18 @@
   "author": "Kong <office@konghq.com>",
   "main": "src/main.min.js",
   "scripts": {
-    "bootstrap": "npm run build:sr",
-    "prebuild": "npm run clean",
-    "build": "npm run generate:ca-certs && npm run build:sr && npm run build:app",
-    "build:app": "esr --cache ./scripts/build.ts --noErrorTruncation",
-    "build:main.min.js": "cross-env NODE_ENV=development esr esbuild.main.ts",
-    "build:sr": "npm run generate:ca-certs && esr esbuild.sr.ts",
-    "bump-version": "esr scripts/bumpVersion.ts",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "postclean": "rimraf build dist src/main.min.js",
-    "generate:ca-certs": "esr scripts/generateCACerts.ts",
-    "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "package": "cross-env USE_HARD_LINKS=false electron-builder build --config electron-builder.config.js",
     "start": "concurrently -n dev,app --kill-others \"npm run start:dev-server\" \"npm run start:electron\"",
     "start:dev-server": "vite dev",
-    "start:electron": "cross-env NODE_ENV=development esr esbuild.main.ts && electron .",
+    "start:electron": "npm run build:main && electron .",
     "test": "jest",
-    "test:watch": "jest --watch",
     "type-check": "tsc --noEmit -p tsconfig.build.json",
-    "type-check:watch": "npm run type-check -- --watch"
+    "clean": "rimraf dist build src/main.min.js *.tsbuildinfo",
+    "build:main": "esr esbuild.main.ts",
+    "build:prod": "esr --cache ./scripts/build.ts --noErrorTruncation",
+    "package": "esr --cache ./scripts/build.ts --noErrorTruncation && esr scripts/package.ts",
+    "build:sr": "esr scripts/generateCACerts.ts && esr esbuild.sr.ts",
+    "bump-version": "esr scripts/bumpVersion.ts",
+    "lint": "eslint . --ext .js,.ts,.tsx --cache"
   },
   "dependencies": {
     "@getinsomnia/node-libcurl": "2.3.5-1",
