+++ packages/openapi-2-kong/package.json
@@ -6,14 +6,10 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "bootstrap": "npm run build",
+    "bootstrap": "tsc --build tsconfig.build.json",
     "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "postclean": "rimraf dist",
-    "build": "tsc --build tsconfig.build.json",
-    "test": "jest",
-    "prepublish": "npm run build"
+    "clean": "rimraf dist *.tsbuildinfo",
+    "test": "jest"
   },
   "dependencies": {
     "openapi-types": "^8.0.0",
